added periodic jobs, sql_script extended, quartz integrated

+++ Consumer_module/src/main/java/cheboksarov/blps_lab3/dto/DoBetDto.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class DoBetDto {
     Long matchId;
     String event;
-    Integer bet;
+    Double bet;
 }
