+++ Consumer_module/src/main/java/cheboksarov/blps_lab3/dto/DoBetRequest.java
@@ -13,6 +13,6 @@ import lombok.NoArgsConstructor;
 public class DoBetRequest {
     Long matchId;
     String event;
-    Integer bet;
+    Double bet;
     Long credentialId;
 }
