+++ Producer_module/src/main/java/cheboksarov/blps_lab3/dto/DoBetRequest.java
@@ -9,6 +9,6 @@ import lombok.Data;
 public class DoBetRequest {
     Long matchId;
     String event;
-    Integer bet;
+    Double bet;
     Long credentialId;
 }
